+++ app/index.html
@@ -10,7 +10,7 @@
     {{content-for "head"}}
 
     <link integrity="" rel="stylesheet" href="{{rootURL}}assets/vendor.css">
-    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/coolstore.css">
+    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/coolwriter.css">
 
     {{content-for "head-footer"}}
   </head>
@@ -18,7 +18,7 @@
     {{content-for "body"}}
 
     <script src="{{rootURL}}assets/vendor.js"></script>
-    <script src="{{rootURL}}assets/coolstore.js"></script>
+    <script src="{{rootURL}}assets/coolwriter.js"></script>
     <!--script src="vendor/mark.js/dist/mark.min.js"></script-->
 
     {{content-for "body-footer"}}
