+++ app/styles/app.sass
@@ -2,12 +2,15 @@
 
 body
   background-color: lightgrey
-  margin-top: 65px
 
 mark
   background-color: lightblue
   color: white
   text-shadow: -3px -3px 0 #000
+
+main
+  margin-left: 20px
+  margin-left: 20px
 
 .welcomeText
   font-family: Verdana
