+++ app/templates/books.hbs
@@ -1,6 +1,6 @@
 
 {{#link-to 'books.new'}}
-  {{#paper-button raised=true}}{{paper-icon 'book'}} {{t 'book.new'}}{{/paper-button}}
+  {{#paper-button raised=true}}{{paper-icon 'book'}}{{t 'book.new'}}{{/paper-button}}
 {{/link-to}}
 <br>
 <br>
