+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "coolstore",
+  "name": "coolwriter",
   "version": "0.0.0",
   "private": true,
-  "description": "Small description for coolstore goes here",
+  "description": "Small description for coolwriter goes here",
   "license": "MIT",
   "author": "",
   "directories": {
@@ -17,43 +17,42 @@
     "test": "ember test"
   },
   "devDependencies": {
-    "broccoli-asset-rev": "^2.4.5",
-    "ember-ajax": "^3.0.0",
-    "ember-auto-import": "^1.0.1",
-    "ember-cli": "^3.1.4",
-    "ember-cli-app-version": "^3.0.0",
-    "ember-cli-babel": "^6.6.0",
-    "ember-cli-dependency-checker": "^2.0.0",
-    "ember-cli-eslint": "^4.2.1",
-    "ember-cli-htmlbars": "^2.0.1",
-    "ember-cli-htmlbars-inline-precompile": "^1.0.0",
-    "ember-cli-inject-live-reload": "^1.4.1",
-    "ember-cli-qunit": "^4.1.1",
+    "broccoli-asset-rev": "^2.7.0",
+    "ember-ajax": "^3.1.0",
+    "ember-auto-import": "^1.2.2",
+    "ember-cli": "^3.2.0",
+    "ember-cli-app-version": "^3.2.0",
+    "ember-cli-babel": "^6.14.1",
+    "ember-cli-dependency-checker": "^2.2.0",
+    "ember-cli-eslint": "^4.2.3",
+    "ember-cli-htmlbars": "^2.0.3",
+    "ember-cli-htmlbars-inline-precompile": "^1.0.3",
+    "ember-cli-inject-live-reload": "^1.7.0",
+    "ember-cli-qunit": "^4.3.2",
     "ember-cli-sass": "^7.1.7",
     "ember-cli-shims": "^1.2.0",
-    "ember-cli-sri": "^2.1.0",
-    "ember-cli-uglify": "^2.0.0",
-    "ember-data": "~3.1.0",
+    "ember-cli-sri": "^2.1.1",
+    "ember-cli-uglify": "^2.1.0",
+    "ember-data": "~3.2.0",
     "ember-emberfire-find-query": "^1.0.3",
     "ember-export-application-global": "^2.0.0",
     "ember-i18n": "^5.2.1",
-    "ember-load-initializers": "^1.0.0",
+    "ember-load-initializers": "^1.1.0",
     "ember-maybe-import-regenerator": "^0.1.6",
     "ember-paper": "^1.0.0-beta.7",
-    "ember-popper": "^0.9.0",
-    "ember-resolver": "^4.0.0",
-    "ember-source": "~3.1.0",
-    "ember-welcome-page": "^3.0.0",
+    "ember-popper": "^0.9.1",
+    "ember-resolver": "^4.5.6",
+    "ember-source": "~3.2.2",
+    "ember-welcome-page": "^3.2.0",
     "emberfire": "^2.0.10",
-    "eslint-plugin-ember": "^5.0.0",
+    "eslint-plugin-ember": "^5.2.0",
     "eslint-plugin-ember-suave": "^1.0.0",
-    "loader.js": "^4.2.3",
+    "loader.js": "^4.7.0",
     "mark.js": "^8.11.1",
     "torii": "^0.10.1"
   },
   "dependencies": {
-    "bootstrap": "^4.1.2",
-    "ember-template-lint": "^0.8.20",
+    "ember-template-lint": "^0.8.22",
     "jquery": "^3.3.1",
     "popper.js": "^1.14.3"
   }
